+++ services/graphite/config_test.go
@@ -1,17 +1,16 @@
-package httpd_test
+package graphite_test
 
 import (
 	"testing"
 
 	"github.com/BurntSushi/toml"
-	"github.com/influxdb/influxdb/httpd"
+	"github.com/influxdb/influxdb/services/httpd"
 )
 
 func TestConfig_Parse(t *testing.T) {
 	// Parse configuration.
 	var c httpd.Config
 	if _, err := toml.Decode(`
-enabled = true
 bind-address = ":8080"
 auth-enabled = true
 log-enabled = true
@@ -22,9 +21,7 @@ pprof-enabled = true
 	}
 
 	// Validate configuration.
-	if c.Enabled != true {
-		t.Fatalf("unexpected enabled: %v", c.Enabled)
-	} else if c.BindAddress != ":8080" {
+	if c.BindAddress != ":8080" {
 		t.Fatalf("unexpected bind address: %s", c.BindAddress)
 	} else if c.AuthEnabled != true {
 		t.Fatalf("unexpected auth enabled: %v", c.AuthEnabled)
